Refactor pattern renaming functionality by introducing a new PatternRenamingTool component. This change simplifies the ModalContentParticipantModification and ModalContentStimulusModification components by offloading pattern renaming logic and UI elements, enhancing code maintainability and user experience.
